feat(type_inference): complete type-aware method completion and CFG integration
This update finalizes the implementation of type-aware method completions based on inferred types from the Control Flow Graph (CFG). It enhances the completion system to provide method suggestions that are contextually aware of the receiver's type, including support for both method-level and top-level code. Additionally, the integration with RBS for built-in method signatures has been completed, improving the accuracy and relevance of method completions. Comprehensive tests have been added to validate these enhancements, ensuring a better developer experience.

+++ src/handlers/notification.rs
@@ -33,7 +33,7 @@ pub async fn handle_initialize(
                 "Parent process ID received: {}. Starting process monitor.",
                 process_id
             );
-            lang_server.set_parent_process_id(Some(process_id as u32));
+            lang_server.set_parent_process_id(Some(process_id));
         } else {
             info!(
                 "Invalid parent process ID received ({}), skipping process monitoring",
